refactor: move `publicUserAccountId` to backend utils
