+++ apps/hash-api/src/auth/create-auth-handlers.ts
@@ -1,6 +1,7 @@
 import { getRequiredEnv } from "@local/hash-backend-utils/environment";
 import { getHashInstance } from "@local/hash-backend-utils/hash-instance";
 import type { Logger } from "@local/hash-backend-utils/logger";
+import { publicUserAccountId } from "@local/hash-backend-utils/public-user-account-id";
 import type { Session } from "@ory/client";
 import type { AxiosError } from "axios";
 import type { Express, Request, RequestHandler } from "express";
@@ -15,7 +16,6 @@ import { systemAccountId } from "../graph/system-account";
 import { hydraAdmin } from "./ory-hydra";
 import type { KratosUserIdentity } from "./ory-kratos";
 import { kratosFrontendApi } from "./ory-kratos";
-import { publicUserAccountId } from "./public-user-account-id";
 
 const KRATOS_API_KEY = getRequiredEnv("KRATOS_API_KEY");
