+++ apps/hash-api/src/auth/get-actor-id.ts
@@ -1,6 +1,5 @@
+import { publicUserAccountId } from "@local/hash-backend-utils/public-user-account-id";
 import type { Request } from "express";
-
-import { publicUserAccountId } from "./public-user-account-id";
 
 export const getActorIdFromRequest = (request: Request) =>
   request.user?.accountId ?? publicUserAccountId;
