+++ apps/hash-api/src/generate-ontology-type-ids.ts
@@ -6,6 +6,7 @@ import { fileURLToPath } from "node:url";
 import { createGraphClient } from "@local/hash-backend-utils/create-graph-client";
 import { getRequiredEnv } from "@local/hash-backend-utils/environment";
 import { Logger } from "@local/hash-backend-utils/logger";
+import { publicUserAccountId } from "@local/hash-backend-utils/public-user-account-id";
 import { currentTimeInstantTemporalAxes } from "@local/hash-isomorphic-utils/graph-queries";
 import type {
   DataTypeWithMetadata,
@@ -14,7 +15,6 @@ import type {
 } from "@local/hash-subgraph";
 import { extractBaseUrl } from "@local/hash-subgraph/type-system-patch";
 
-import { publicUserAccountId } from "./auth/public-user-account-id";
 import type {
   ImpureGraphContext,
   ImpureGraphFunction,
